Improvements on host network update hosts status
